refactor : 요구사항에 따른 변경 연관관계 설정으로 인한 수정 (User table 생성)
